refact(Events) : Event들을 다른 scripts에서 관리
