+++ Assets/Scripts/Event/Events.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: bacaedddef8cb43e4bf856d410b469a3
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
